use new queue implementation
